separate the parser from the driver and shift file itself, and create a parser controller to handle the parsing of the files
